Factor RouterScheduler out of monad-mock-swarm
This allows monad-quic to no longer depend on monad-mock-swarm.
