+++ monad-mock-swarm/tests/many_nodes.rs
@@ -5,11 +5,10 @@ use common::QuicSwarm;
 use monad_consensus_types::transaction_validator::MockValidator;
 use monad_gossip::mock::MockGossipConfig;
 use monad_mock_swarm::{
-    mock::{MockMempoolConfig, NoSerRouterConfig},
-    mock_swarm::UntilTerminator,
-    swarm_relation::NoSerSwarm,
+    mock::MockMempoolConfig, mock_swarm::UntilTerminator, swarm_relation::NoSerSwarm,
 };
 use monad_quic::QuicRouterSchedulerConfig;
+use monad_router_scheduler::NoSerRouterConfig;
 use monad_testutil::swarm::{create_and_run_nodes, SwarmTestConfig};
 use monad_transformer::{BwTransformer, BytesTransformer, GenericTransformer, LatencyTransformer};
 use monad_wal::mock::MockWALoggerConfig;
