+++ monad-quic/Cargo.toml
@@ -14,8 +14,8 @@ bench = false
 monad-crypto = { path = "../monad-crypto", features = ["rustls"] }
 monad-executor-glue = { path = "../monad-executor-glue" }
 monad-gossip = { path = "../monad-gossip" }
-monad-mock-swarm = { path = "../monad-mock-swarm" }
 monad-executor = { path = "../monad-executor" }
+monad-router-scheduler = { path = "../monad-router-scheduler" }
 monad-types = { path = "../monad-types" }
 
 futures = { workspace = true }
