+++ monad-virtual-bench/Cargo.toml
@@ -20,6 +20,7 @@ monad-executor-glue = { path = "../monad-executor-glue" }
 monad-gossip = { path = "../monad-gossip" }
 monad-mock-swarm = { path = "../monad-mock-swarm" }
 monad-quic = { path = "../monad-quic" }
+monad-router-scheduler = { path = "../monad-router-scheduler" }
 monad-state = { path = "../monad-state" }
 monad-testutil = { path = "../monad-testutil" }
 monad-transformer = { path = "../monad-transformer" }
